requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
openai==1.51.0
python-dotenv==1.0.1
